sped up getRateThroughTimeMatrix R code
I realize the goal is to put this all in C, but in studying the R code I was able to speed it up considerably. Bootlenecks were:

1. Vectorized exponentialRate. This was 8X slower than the non-vectorized version.
2. Slow SafeCompare. Called very often, and performed unnecessary calculations.
3. Checks made when they should only be conditionally checked.

getRateThroughTimeMatrix still takes ~67 seconds to process the Jetz example tree, so C is definitely still needed, but it is better than the 40 minute execution time of the original R code.

Okay, off to tackle the C code with the lessons learned above!